Cannot use uint8_t for 4 digit yearsgit status using uint16_t instead

+++ libindi/drivers/telescope/lx200generic.cpp
@@ -792,7 +792,7 @@ bool LX200Generic::Abort()
     return true;
 }
 
-bool LX200Generic::setLocalDate(uint8_t days, uint8_t months, uint8_t years)
+bool LX200Generic::setLocalDate(uint8_t days, uint8_t months, uint16_t years)
 {
     return (setCalenderDate(PortFD, days, months, years) == 0);
 }
